+++ ssl-proxy.sh
@@ -7,7 +7,7 @@ fi
 
 while true
 do
-  ./ssl-proxy-linux-amd64 -from 0.0.0.0:443 -to 0.0.0.0:8501 -redirectHTTP -domain=repscoreportal.org &
+  ./ssl-proxy-linux-amd64 -from https://0.0.0.0:443 -to http://0.0.0.0:8501 -redirectHTTP -domain=repscoreportal.org &
   kill $( jobs -p )
   sleep 7776000  # sleep for 90 days before regenerating new SSL certificates
 done